requirement.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
certifi==2025.10.5
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.3
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-multipart==0.0.20
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
